app/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
Jinja2==3.1.6
## The following requirements were added by pip freeze:
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.8.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==1.35.0
nest-asyncio==1.6.0
numba==0.61.2
numpy==1.24.4
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyzmq==26.4.0
rdkit==2024.9.6
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.1.3
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SpectraFP==1.2.4
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
streamlit==1.44.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13